+++ lib/plugins/config.js
@@ -2,11 +2,11 @@
 
 const BbPromise = require('bluebird');
 const config = require('@serverless/utils/config');
-const isTabTabCompletionSupported = require('../../utils/tabCompletion/isSupported');
-const muteConsoleLog = require('../../utils/log/muteConsoleLog');
-const tabtabOptions = require('../../utils/tabCompletion/tabtabOptions');
-const isNpmPackageWritable = require('../../utils/npmPackage/isWritable');
-const isNpmGlobalPackage = require('../../utils/npmPackage/isGlobal');
+const isTabTabCompletionSupported = require('../utils/tabCompletion/isSupported');
+const muteConsoleLog = require('../utils/log/muteConsoleLog');
+const tabtabOptions = require('../utils/tabCompletion/tabtabOptions');
+const isNpmPackageWritable = require('../utils/npmPackage/isWritable');
+const isNpmGlobalPackage = require('../utils/npmPackage/isGlobal');
 
 // class wide constants
 const validProviders = new Set(['aws']);
