+++ lib/plugins/index.js
@@ -2,7 +2,7 @@
 
 module.exports = [
   require('./interactiveCli'),
-  require('./config/config.js'),
+  require('./config.js'),
   require('./create/create.js'),
   require('./install/install.js'),
   require('./package/package.js'),
